refactor: enums in to entity pack
