+++ src/main/java/com/paymentteamproject/domain/payment/entity/Payment.java
@@ -1,7 +1,6 @@
 package com.paymentteamproject.domain.payment.entity;
 
 import com.paymentteamproject.common.entity.BaseEntity;
-import com.paymentteamproject.domain.payment.enums.PaymentStatus;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
