+++ src/main/java/com/paymentteamproject/domain/payment/entity/PaymentStatus.java
@@ -1,4 +1,4 @@
-package com.paymentteamproject.domain.payment.enums;
+package com.paymentteamproject.domain.payment.entity;
 
 public enum PaymentStatus {
     // 대기, 성공, 실패, 환불
