+++ src/main/java/com/paymentteamproject/domain/refund/entity/RefundStatus.java
@@ -1,4 +1,4 @@
-package com.paymentteamproject.domain.refund.enums;
+package com.paymentteamproject.domain.refund.entity;
 
 public enum RefundStatus {
     // 요청, 성공, 실패
